Move chai.py into lib folder.
